Crawler/LEAGUE-V4/pyproject.toml
[tool.poetry]
name = "accoun-v1"
version = "0.1.0"
description = ""
authors = ["YoungHyeon1 <73692488+YoungHyeon1@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
boto3 = "^1.34.83"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
